+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation/include/camera_avfoundation/FLTCam_Test.h
@@ -40,6 +40,9 @@ typedef id<FLTCaptureDeviceControlling> (^CaptureDeviceFactory)(void);
 /// True when images from the camera are being streamed.
 @property(assign, nonatomic) BOOL isStreamingImages;
 
+@property(readonly, nonatomic) AVCaptureSession *videoCaptureSession;
+@property(readonly, nonatomic) AVCaptureSession *audioCaptureSession;
+
 /// A dictionary to retain all in-progress FLTSavePhotoDelegates. The key of the dictionary is the
 /// AVCapturePhotoSettings's uniqueID for each photo capture operation, and the value is the
 /// FLTSavePhotoDelegate that handles the result of each photo capture operation. Note that photo